chore: Ajustar campos de float para double no BD e adicionar recuperação dos dados ao lançar uma exceção de validação

+++ database/migrations/2022_12_29_142121_create_fatura_pedido.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('fatura_pedidos', function (Blueprint $table) {
             $table->id();
-            $table->float('valor');
+            $table->double('valor');
             $table->date('vencimento');
             $table->unsignedBigInteger('venda_id');
             $table->foreign('venda_id')->references('id')->on('pedidos')->onDelete('cascade');
